+++ voicedna/__init__.py
@@ -1,4 +1,5 @@
 from voice_dna import VoiceDNA
+from .framework import VoiceDNAProcessor
 
 from .plugins import (
     Base64PassThroughFilter,
@@ -9,6 +10,7 @@ from .plugins import (
 
 __all__ = [
     "VoiceDNA",
+    "VoiceDNAProcessor",
     "IVoiceDNAFilter",
     "PluginManager",
     "Base64PassThroughFilter",